+++ jetpack_vendor/i18n-map.php
@@ -44,5 +44,9 @@ return array(
       'path' => 'jetpack_vendor/automattic/jetpack-sync',
       'ver' => '1.37.9999999-dev',
     ),
+    'jetpack-videopress-pkg' => array(
+      'path' => 'jetpack_vendor/automattic/jetpack-videopress',
+      'ver' => '0.1.9999999-dev',
+    ),
   ),
 );